fix: #103 the preview will now use the default configuration when called without arguments. the lighthouse crawler will no longer try to send the default configuration but rather rely on the preview taking care of that anyway.
