+++ backend/data-service/src/main/java/it/bz/opendatahub/webcomponents/dataservice/application/port/in/GetDefaultPreviewSnippetUseCase.java
@@ -1,0 +1,7 @@
+package it.bz.opendatahub.webcomponents.dataservice.application.port.in;
+
+import lombok.NonNull;
+
+public interface GetDefaultPreviewSnippetUseCase {
+	String getDefaultPreviewSnippet(@NonNull String webcomponentUuid, @NonNull String versionTag);
+}
